Removed a lot of logic that is relevant to ChatServer
A lot of changes to remove logic that isn't actually CoreServer but more likely to be used by ChatServer.
End result of this repo is to make a server that can be reused.
Some stuff like command invokation should probably be updated because, you need to basically change MessageProcessor and ServerCommand interfaces.